Making artisan fresh more secure

+++ src/Illuminate/Foundation/Console/FreshCommand.php
@@ -28,6 +28,12 @@ class FreshCommand extends Command {
 	{
 		$files = new Filesystem;
 
+		if($files->exists($this->laravel->storagePath().'/framework/fresh'))
+		{
+			$this->error('You shouldn\'t run this twice, it\'s dangerous... Remove [/storage/framework/fresh] and try again if you\'re sure.');
+			return;
+		}
+
 		$files->deleteDirectory(app_path('Services'));
 		$files->delete(base_path('resources/views/app.blade.php'));
 		$files->delete(base_path('resources/views/home.blade.php'));
@@ -42,6 +48,8 @@ class FreshCommand extends Command {
 		$files->deleteDirectory(base_path('resources/assets/less/bootstrap'));
 
 		$files->put(app_path('Http/routes.php'), $files->get(__DIR__.'/stubs/fresh-routes.stub'));
+
+		touch($this->laravel->storagePath().'/framework/fresh');
 
 		$this->info('Scaffolding removed! Enjoy your fresh start.');
 	}
